+++ CHANGELOG.md
@@ -25,6 +25,7 @@ from pydantic import validate_arguments
 ## 2.2.8
 
 - flat routes with no more zz_conn
+- added from pp.import_gds import add_ports_from_markers_square
 
 ## 2.2.7